% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_A}
\alias{Generate_A}
\title{A matrix entries generating function}
\usage{
Generate_A(X, Y, A, i, j, Sigma_Inv, N, p, B, gamma, tau, nu_1, prop_var1)
}
\arguments{
\item{X}{n * K matrix input}

\item{Y}{n * p matrix input}

\item{A}{p * p matrix input}

\item{i}{positive integer input in between 1 and p}

\item{j}{positive integer input in between 1 and p}

\item{Sigma_Inv}{p * p matrix input}

\item{N}{positive integer input}

\item{p}{positive integer input}

\item{B}{p * K matrix input}

\item{gamma}{scalar input, 0 or 1}

\item{tau}{postive scalar input}

\item{nu_1}{postive scalar input}

\item{prop_var1}{postive scalar input}
}
\value{
scalar a value
}
\description{
A matrix entries generating function
}
