% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Agamma}
\alias{Generate_Agamma}
\title{A matrix entries and gamma generating function}
\usage{
Generate_Agamma(
  X,
  Y,
  A,
  i,
  j,
  Sigma_Inv,
  N,
  p,
  B,
  gamma,
  tau,
  rho,
  nu_1,
  prop_var1,
  prop_var2
)
}
\arguments{
\item{X}{n * K matrix input}

\item{Y}{n * p matrix input}

\item{A}{p * p matrix input}

\item{i}{integer input in between 1 and p}

\item{j}{integer input in between 1 and p}

\item{Sigma_Inv}{p * 1 vector input}

\item{N}{positive scalar input}

\item{p}{positive integer input}

\item{B}{p * K matrix input}

\item{gamma}{scalar input, 0 or 1}

\item{tau}{positive scalar input}

\item{rho}{scalar input in between 0 and 1}

\item{nu_1}{positive scalar input}

\item{prop_var1}{postive scalar input}

\item{prop_var2}{postive scalar input}
}
\value{
scalar a value and scalar gamma value, 0 or 1
}
\description{
A matrix entries and gamma generating function
}
