% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_B}
\alias{Generate_B}
\title{B matrix entries generating function}
\usage{
Generate_B(xz, b_vec, b, x_mat, x_vec, sigma, eta, phi, nu2)
}
\arguments{
\item{xz}{scalar input}

\item{b_vec}{vector input}

\item{b}{scalar input}

\item{x_mat}{matrix input}

\item{x_vec}{vector input}

\item{sigma}{positive scalar input}

\item{eta}{positive scalar input}

\item{phi}{scalar input 0 or 1}

\item{nu2}{positive scalar input}
}
\value{
scalar value generated from normal distribution
}
\description{
B matrix entries generating function
}
