% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Bphi}
\alias{Generate_Bphi}
\title{B matrix entries and phi generating function}
\usage{
Generate_Bphi(
  X,
  Y,
  B,
  i,
  j,
  Sigma_Inv,
  Mult_Inv_Y,
  phi,
  eta,
  psi,
  nu_2,
  prop_var2
)
}
\arguments{
\item{X}{n * K matrix input}

\item{Y}{n * p matrix input}

\item{B}{p * K matrix input}

\item{i}{positive integer input in between 1 and p}

\item{j}{positive integer input in between 1 and K}

\item{Sigma_Inv}{p * p matrix input}

\item{Mult_Inv_Y}{p * n matrix input}

\item{phi}{scalar input 0 or 1}

\item{eta}{positive scalar input}

\item{psi}{scalar input in between 0 and 1}

\item{nu_2}{positive scalar input}

\item{prop_var2}{positive scalar input}
}
\value{
scalar b value and scalar phi value, 0 or 1
}
\description{
B matrix entries and phi generating function
}
