% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Eta}
\alias{Generate_Eta}
\title{Eta generating function}
\usage{
Generate_Eta(b, phi, a_eta, b_eta, nu_2)
}
\arguments{
\item{b}{scalar input}

\item{phi}{scalar input either 0 or 1}

\item{a_eta}{positive scalar input, first parameter of inverse gamma distribution}

\item{b_eta}{positive scalar input, second parameter of inverse gamma distribution}

\item{nu_2}{positive scalar input generated from inverse gamma distribution}
}
\value{
scalar value generated from inverse gamma distribution, phi / rgamma(1, a_eta + 1/2, b^2/2 + b_eta) + (1 - phi) / rgamma(1, a_eta + 1/2, b^2/(2 * nu_2) + b_eta)
}
\description{
Eta generating function
}
