% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Gamma}
\alias{Generate_Gamma}
\title{Gamma generating function}
\usage{
Generate_Gamma(a, rho, tau, nu_1)
}
\arguments{
\item{a}{scalar input}

\item{rho}{scalar input between 0 and 1}

\item{tau}{positive scalar input}

\item{nu_1}{positive scalar input}
}
\value{
scalar value generated from a bernoulli distribution i.e. 0 or 1
}
\description{
Gamma generating function
}
