% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Phi}
\alias{Generate_Phi}
\title{Phi generating function}
\usage{
Generate_Phi(b, psi, eta, nu_2)
}
\arguments{
\item{b}{scalar input}

\item{psi}{scalar input between 0 and 1}

\item{eta}{positive scalar input}

\item{nu_2}{positive scalar input}
}
\value{
scalar value generated from a bernoulli distribution i.e. 0 or 1
}
\description{
Phi generating function
}
