% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Psi}
\alias{Generate_Psi}
\title{Psi generating function}
\usage{
Generate_Psi(Phi, d, a_psi, b_psi)
}
\arguments{
\item{Phi}{matrix input}

\item{d}{non-negative scalar input}

\item{a_psi}{positive scalar input}

\item{b_psi}{positive scalar input}
}
\value{
scalar value generated from beta distribution, rbeta(1, sum(Phi) + a_psi, d - sum(Phi) + b_psi)
}
\description{
Psi generating function
}
