% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Sigma}
\alias{Generate_Sigma}
\title{Sigma generating function}
\usage{
Generate_Sigma(n, z_sum, a_sigma, b_sigma)
}
\arguments{
\item{n}{positive integer input}

\item{z_sum}{non-negative scalar input}

\item{a_sigma}{positive scalar input}

\item{b_sigma}{positive scalar input}
}
\value{
scalar value generated from inverse gamma distribution, 1 / rgamma(1, n/2 + a_sigma, z_sum/2 + b_sigma)
}
\description{
Sigma generating function
}
