% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Tau}
\alias{Generate_Tau}
\title{Tau generating function}
\usage{
Generate_Tau(a, gamma, a_tau, b_tau, nu_1)
}
\arguments{
\item{a}{scalar input}

\item{gamma}{scalar input either 0 or 1}

\item{a_tau}{positive scalar input}

\item{b_tau}{positive scalar input}

\item{nu_1}{positive scalar input}
}
\value{
scalar value generated from inverse gamma distribution, gamma / rgamma(1, a_tau + 1/2, a^2/2 + b_tau) + (1 - gamma) / rgamma(1, a_tau + 1/2, a^2/(2 * nu_1) + b_tau)
}
\description{
Tau generating function
}
