% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{LL}
\alias{LL}
\title{Log-likelihood generating function}
\usage{
LL(A, B, X, Y, Sigma_Inv, p, N)
}
\arguments{
\item{A}{p * p matrix input}

\item{B}{p * K matrix input}

\item{X}{n * K matrix input}

\item{Y}{n * p matrix input}

\item{Sigma_Inv}{p * 1 vector input}

\item{p}{positive integer input}

\item{N}{positive scalar input}
}
\value{
Log likelihood value corresponding to a particular A, B, X, Y and Sigma_Inv
}
\description{
Log-likelihood generating function
}
