% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Target_A}
\alias{Target_A}
\title{Target function for a particular A entry}
\usage{
Target_A(X, Y, A, a, N, Sigma_Inv, p, B, gamma, tau, nu_1)
}
\arguments{
\item{X}{n * K matrix input}

\item{Y}{n * p matrix input}

\item{A}{p * p matrix input}

\item{a}{scalar input}

\item{N}{positive scalar input}

\item{Sigma_Inv}{p * p matrix input}

\item{p}{positive integer input}

\item{B}{p * K matrix input}

\item{gamma}{scalar input, 0 or 1}

\item{tau}{positive scalar input}

\item{nu_1}{positive scalar input}
}
\value{
scalar target value
}
\description{
Target function for a particular A entry
}
