% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Target_Agamma}
\alias{Target_Agamma}
\title{Target function for a particular A entry and corresponding gamma}
\usage{
Target_Agamma(X, Y, A, a, N, Sigma_Inv, p, B, gamma, tau, rho, nu_1)
}
\arguments{
\item{X}{n * K matrix input}

\item{Y}{n * p matrix input}

\item{A}{p * p matrix input}

\item{a}{scalar input}

\item{N}{positive scalar input}

\item{Sigma_Inv}{p * 1 vector input}

\item{p}{positive integer input}

\item{B}{p * K matrix input}

\item{gamma}{scalar input, 0 or 1}

\item{tau}{positive scalar input}

\item{rho}{scalar input in between 0 and 1}

\item{nu_1}{positive scalar input}
}
\value{
scalar target value corresponding to a particular a and gamma value
}
\description{
Target function for a particular A entry and corresponding gamma
}
