% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Target_Bphi}
\alias{Target_Bphi}
\title{Target function for a particular A entry and corresponding gamma}
\usage{
Target_Bphi(X, Y, B, Sigma_Inv, MultMat_Y, b, phi, eta, psi, nu_2)
}
\arguments{
\item{X}{n * K matrix input}

\item{Y}{n * p matrix input}

\item{B}{p * K matrix input}

\item{Sigma_Inv}{p * 1 vector input}

\item{MultMat_Y}{p * n matrix input}

\item{b}{scalar input}

\item{phi}{scalar input, 0 or 1}

\item{eta}{positive scalar input}

\item{psi}{scalar input in between 0 and 1}

\item{nu_2}{positive scalar input}
}
\value{
scalar target value
}
\description{
Target function for a particular A entry and corresponding gamma
}
