% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.R
\name{Generate_Rho}
\alias{Generate_Rho}
\title{Rho generating function}
\usage{
Generate_Rho(Gamma, p, a_rho, b_rho)
}
\arguments{
\item{Gamma}{matrix input}

\item{p}{positive scalar input}

\item{a_rho}{positive scalar input}

\item{b_rho}{positive scalar input}
}
\value{
scalar value generated from beta distribution, rbeta(1, sum(Gamma) + a_rho, p * (p - 1) - sum(Gamma) + b_rho)
}
\description{
Rho generating function
}
