% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGM.R
\name{RGM}
\alias{RGM}
\title{Title Fitting Reciprocal Graphical Models for Integrative Gene Regulatory Network}
\usage{
RGM(
  X,
  Y,
  A0 = NULL,
  B0 = NULL,
  D = NULL,
  a_tau = 0.1,
  b_tau = 0.1,
  a_rho = 0.5,
  b_rho = 0.5,
  nu_1 = 1e-04,
  a_eta = 0.1,
  b_eta = 0.1,
  a_psi = 0.5,
  b_psi = 0.5,
  nu_2 = 1e-04,
  a_sigma = 0.1,
  b_sigma = 0.1,
  Prop_varA = 0.1,
  Prop_VarB = 0.1,
  niter = 10000
)
}
\arguments{
\item{X}{n * k matrix data input, each row denotes a particular observation and each column denotes a particular DNA.}

\item{Y}{n * p matrix data input, each row denotes a particular observation and each column denotes a particular gene.}

\item{A0}{p * p matrix input. It is the initial value of the gene gene interaction matrix. Random initialization of this matrix may take very large iterations for the chain to converge. Default value is NULL  .}

\item{B0}{p * k matrix input. It is the initial value of the gene DNA interaction matrix. Each row corresponds to a particular gene, each column corresponds to a particular DNA. Random initialization of this matrix may take long time for the chain to converge. Default value is NULL.}

\item{D}{p * k indicator matrix input. Each row corresponds to a particular gene, each column corresponds to a particular DNA. An entry is 1 if the corresponding DNA can effect the corresponding gene, else 0. Default value is NULL. If not given, all entry of D is taken to be 1.}

\item{a_tau}{positive scalar input. It corresponds to the first parameter of a Inverse Gamma distribution. Default value is 0.1.}

\item{b_tau}{positive scalar input. It corresponds to the second parameter of a Inverse Gamma distribution. Default value is 0.1.}

\item{a_rho}{positive scalar input. It corresponds to the first parameter of a Beta distribution. Default value is 0.5.}

\item{b_rho}{positive scalar input. It corresponds to the first parameter of a Beta distribution. Default value is 0.5.}

\item{nu_1}{positive scalar input. It corresponds to the variance of slab part of of spike and slab distribution of A. Default value is 0.0001.}

\item{a_eta}{positive scalar input. It corresponds to the first parameter of a Inverse Gamma distribution. Default value is 0.1.}

\item{b_eta}{positive scalar input. It corresponds to the second parameter of a Inverse Gamma distribution. Default value is 0.1.}

\item{a_psi}{positive scalar input. It corresponds to the first parameter of a Beta distribution. Default value is 0.5.}

\item{b_psi}{positive scalar input. It corresponds to the first parameter of a Beta distribution. Default value is 0.5.}

\item{nu_2}{positive scalar input. It corresponds to the variance of slab part of of spike and slab distribution of B. Default value is 0.0001.}

\item{a_sigma}{postive scalar input. It corresponds to the first parameter of a Inverse Gamma distribution. Default value is 0.1.}

\item{b_sigma}{postive scalar input. It corresponds to the second parameter of a Inverse Gamma distribution. Default value is 0.1.}

\item{Prop_varA}{postive scalar input. It corresponds to the variance of the normal distribution of proposing A matrix terms. Default value is 0.1.}

\item{Prop_VarB}{postive scalar input. It corresponds to the variance of the normal distribution of proposing B matrix terms. Default value is 0.1.}

\item{niter}{positive integer input. It corresponds to number of iterations the markov chain runs. Give niter as large as possible to get the convergence. Minimum value is 10,000. Default value is 10,000.}
}
\value{
A list which will hold A, B, LL:

\item{A}{p * p matrix output of Gene-Gene interactions.}
\item{B}{p * k matrix output of Gene-DNA interactions.}
\item{LL}{niter * 1 vector output of log-likelihood values of the model for all the iterations.}
}
\description{
RGM can be used to fit Reciprocal Graphical Models on gene expression data and DNA level measurements to find the relationship between different genes and the relationship between genes and DNA.
}
\examples{

# -----------------------------------------------------------------
# Example 1:


n = 500
p = 3
k = 3

A = matrix(sample(c(-3,3), p^2, replace = T), p, p)
diag(A) = 0

B = matrix(0, p, k)

for(i in 1:p){

  B[i, i] = sample(c(-3, 3), 1, replace = T)

}

D = diag(3)

Sigma = 0.5 * diag(p)

Mult_Mat = solve(diag(p) - A)

Variance = Mult_Mat \%*\% Sigma \%*\% t(Mult_Mat)

X = matrix(runif(n * k, 0, 5), nrow = n, ncol = k)

Y = matrix(0, nrow = n, ncol = p)

for (i in 1:n) {

 Y[i, ] = mvrnorm(n = 1, Mult_Mat \%*\% B \%*\% X[i, ], Variance)

}

Out = RGM(X, Y, A = A, B = B, D = D)

A = Out$A
B = Out$B
LL = Out$LL
plot(LL, type = 'l', xlab = "Number of Iterations", ylab = "Log-likelihood")


}
